python_requirements.txt
pytest
Pillow
pyscreenshot
python-libxdo
pynput
